+++ .gitignore
@@ -3,7 +3,7 @@
 .env
 npm-debug.log
 yarn-error.log
-/frontend/nova3/node_modules
-/frontend/nova4/node_modules
+/nova3/node_modules
+/nova4/node_modules
 auth.json
 composer.lock
