+++ nova3/resources/css/card.css
