+++ nova3/resources/js/card.js
