+++ nova3/resources/js/components/Card.vue
