+++ nova3/resources/js/components/HeroIconPlus.vue
