+++ nova3/resources/js/flavors/nova.js
