+++ nova4/resources/js/card.js
