+++ nova4/resources/js/components/HeroIconPlus.vue
