+++ nova4/tailwind.config.js
@@ -1,6 +1,6 @@
 // We're going to use the exact Tailwind config that Nova
 // uses, but just change where we look for content.
-let novaTailwindConfig = require('../../vendor/laravel/nova/tailwind.config.js');
+let novaTailwindConfig = require('../vendor/laravel/nova/tailwind.config.js');
 
 novaTailwindConfig.content = [
     './resources/**/*{js,vue}',
