+++ package.json
@@ -3,11 +3,11 @@
     "version": "0.0.0",
     "author": "Aaron Francis <aarondfrancis@gmail.com> (http://aaronfrancis.com/)",
     "scripts": {
-        "watch:nova3": "cd frontend/nova3 && npm run watch",
-        "watch:nova4": "cd frontend/nova4 && npm run watch",
+        "watch:nova3": "cd nova3 && npm run watch",
+        "watch:nova4": "cd nova4 && npm run watch",
         "watch": "concurrently \"npm:watch:*\"",
-        "dev:nova3": "cd frontend/nova3 && npm run dev",
-        "dev:nova4": "cd frontend/nova4 && npm run dev",
+        "dev:nova3": "cd nova3 && npm run dev",
+        "dev:nova4": "cd nova4 && npm run dev",
         "dev": "concurrently \"npm:dev:*\""
     },
     "devDependencies": {
